+++ src/test/office.test.js
@@ -5,14 +5,68 @@ import server from '../server';
 
 chai.should();
 chai.use(chaiHttp);
+let tokenUser;
+let adminToken;
+let idD;
 
+const user = {
+  firstname: 'akaniru',
+  lastname: 'precious',
+  othernames: 'presh',
+  email: 'preshxx@gmail.com',
+  password: '2020ada',
+  type: 'citizen',
+  phonenumber: '+234-7032421061',
+  username: 'preshxxxxxx',
+};
+const user1 = {
+  email: 'preshxx@gmail.com',
+  password: '2020ada',
+};
+const userAdmin = {
+  email: 'example@yahoo.com',
+  password: '2020ada',
+};
+before((done) => {
+  chai.request(server)
+    .post('/api/v1/auth/signup')
+    .send(user)
+    .end(() => {
+      done();
+    });
+});
+before((done) => {
+  chai.request(server)
+    .post('/api/v1/auth/login')
+    .send(user1)
+    .end((err, res) => {
+      tokenUser = res.body.data[0].token;
+      done();
+    });
+});
+before((done) => {
+  chai.request(server)
+    .post('/api/v1/auth/login')
+    .send(userAdmin)
+    .end((err, res) => {
+      adminToken = res.body.data[0].token;
+      done();
+    });
+});
 
 describe('/Post create political office', () => {
   const office = {
     type: 'federal',
-    name: 'presidency',
+    name: 'vice',
     ageLimit: '50',
   };
+
+  const office1 = {
+    type: 'federal',
+    name: 'speaker',
+    ageLimit: '50',
+  };
+
   const noOfficeType = {
     name: 'presidency',
     ageLimit: '50',
@@ -28,7 +82,7 @@ describe('/Post create political office', () => {
 
   const emptyType = {
     type: '   ',
-    name: 'presidency',
+    name: 'vicepresident',
     ageLimit: '50',
   };
   const emptyofficeName = {
@@ -43,12 +97,12 @@ describe('/Post create political office', () => {
   };
   const invalidType = {
     type: 'ioejbermob',
-    name: 'presidency',
+    name: 'cbngov',
     ageLimit: '50',
   };
   const invalidAge = {
-    type: '   ',
-    name: 'presidency',
+    type: 'state',
+    name: 'president',
     ageLimit: '10',
   };
   const invalidAgeType = {
@@ -56,11 +110,14 @@ describe('/Post create political office', () => {
     name: 'councellor',
     ageLimit: 'xx',
   };
+
   it('it should Create a new office with correct status code', (done) => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(office)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
+        idD = res.body.data.id;
         res.should.have.status(201);
         res.body.status.should.be.equal(201);
         done();
@@ -70,11 +127,12 @@ describe('/Post create political office', () => {
   it('it should Create a new office with required fields', (done) => {
     chai.request(server)
       .post('/api/v1/offices')
-      .send(office)
+      .send(office1)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
-        res.body.data[0].should.have.include.key('type');
-        res.body.data[0].should.have.include.key('name');
-        res.body.data[0].should.have.include.key('ageLimit');
+        res.body.data.should.have.include.key('type');
+        res.body.data.should.have.include.key('name');
+        res.body.data.should.have.include.key('agelimit');
         done();
       });
   });
@@ -83,6 +141,7 @@ describe('/Post create political office', () => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(noOfficeType)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
         res.should.have.status(400);
         done();
@@ -92,6 +151,7 @@ describe('/Post create political office', () => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(noOfficeAge)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
         res.should.have.status(400);
         done();
@@ -101,18 +161,20 @@ describe('/Post create political office', () => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(noOfficeName)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
         res.should.have.status(400);
         done();
       });
   });
 
-  it('it should throw a 400 when there are spaces in the party office name field', (done) => {
+  it('it should throw a 406 when there are spaces in the party office name field', (done) => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(emptyofficeName)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
-        res.should.have.status(400);
+        res.should.have.status(406);
         done();
       });
   });
@@ -120,8 +182,9 @@ describe('/Post create political office', () => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(emptyType)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
-        res.should.have.status(406);
+        res.should.have.status(400);
         done();
       });
   });
@@ -130,6 +193,7 @@ describe('/Post create political office', () => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(emptyAge)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
         res.should.have.status(406);
         done();
@@ -140,6 +204,7 @@ describe('/Post create political office', () => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(invalidType)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
         res.should.have.status(406);
         done();
@@ -150,6 +215,7 @@ describe('/Post create political office', () => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(invalidAge)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
         res.should.have.status(406);
         done();
@@ -160,24 +226,13 @@ describe('/Post create political office', () => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(invalidAgeType)
+      .set('x-access-token', adminToken)
       .end((err, res) => {
         res.should.have.status(406);
         done();
       });
   });
 });
-
-describe('GET /offices', () => {
-  it('should get all existing offices', (done) => {
-    chai.request(server)
-      .get('/api/v1/offices')
-      .end((err, res) => {
-        res.should.have.status(200);
-        done();
-      });
-  });
-});
-
 
 describe('GET /offices/:id', () => {
   const office2 = {
@@ -189,13 +244,15 @@ describe('GET /offices/:id', () => {
     chai.request(server)
       .post('/api/v1/offices')
       .send(office2)
+      .set('x-access-token', tokenUser)
       .end(() => {
         done();
       });
   });
   it('should get the matching office', (done) => {
     chai.request(server)
-      .get('/api/v1/offices/1')
+      .get(`/api/v1/offices/${idD}`)
+      .set('x-access-token', tokenUser)
       .end((err, res) => {
         res.should.have.status(200);
         done();
@@ -205,6 +262,7 @@ describe('GET /offices/:id', () => {
   it('should return 406 on nan ids', (done) => {
     chai.request(server)
       .get('/api/v1/offices/sv')
+      .set('x-access-token', tokenUser)
       .end((err, res) => {
         res.should.have.status(406);
         done();
@@ -214,8 +272,21 @@ describe('GET /offices/:id', () => {
   it('should return 404 when id is not found', (done) => {
     chai.request(server)
       .get('/api/v1/offices/10')
+      .set('x-access-token', tokenUser)
       .end((err, res) => {
         res.should.have.status(404);
+        done();
+      });
+  });
+});
+
+describe('GET /offices', () => {
+  it('should get all existing offices', (done) => {
+    chai.request(server)
+      .get('/api/v1/offices')
+      .set('x-access-token', tokenUser)
+      .end((err, res) => {
+        res.should.have.status(200);
         done();
       });
   });
